Added TestCafe test for CreateMeeting
